refactor: establish app and page boundaries
