+++ knip.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://unpkg.com/knip@6/schema.json",
   "entry": [
-    "src/main.tsx",
+    "src/app/main.tsx",
     "public/storybookServiceWorker.js"
   ],
   "includeEntryExports": true
