--- src/features/card/containers/index.ts
@@ -1,8 +1,0 @@
-/**
- * @file Collects the public exports for the features card containers area.
- * Callers can import from this boundary without depending on the area's internal folder layout.
- */
-
-export * from "@/features/card/containers/CardListContainer";
-export * from "@/features/card/containers/CardFormContainer";
-export * from "@/features/card/containers/CardViewContainer";
