--- src/features/deck/containers/index.ts
@@ -1,7 +1,0 @@
-/**
- * @file Collects the public exports for the features deck containers area.
- * Callers can import from this boundary without depending on the area's internal folder layout.
- */
-
-export * from "@/features/deck/containers/DeckListContainer";
-export * from "@/features/deck/containers/DeckFormContainer";
