--- src/features/settings/containers/index.ts
@@ -1,6 +1,0 @@
-/**
- * @file Collects the public exports for the features settings containers area.
- * Callers can import from this boundary without depending on the area's internal folder layout.
- */
-
-export * from "@/features/settings/containers/ConfigContainer";
