--- src/features/study/containers/index.ts
@@ -1,7 +1,0 @@
-/**
- * @file Collects the public exports for the features study containers area.
- * Callers can import from this boundary without depending on the area's internal folder layout.
- */
-
-export * from "@/features/study/containers/DeckStartContainer";
-export * from "@/features/study/containers/DeckSwiperContainer";
