+++ src/features/study/index.ts
@@ -1,0 +1,2 @@
+export { DeckStartContainer } from "./containers/DeckStartContainer";
+export { DeckSwiperContainer } from "./containers/DeckSwiperContainer";
