--- src/page/index.ts
@@ -1,15 +1,0 @@
-/**
- * @file Collects the route-level page exports used by the application router.
- * The router can import every page from one boundary without depending on individual page file
- * paths.
- */
-
-export * from "@/page/DeckList";
-export * from "@/page/DeckFormPage";
-export * from "@/page/DeckStartPage";
-export * from "@/page/DeckSwiperPage";
-export * from "@/page/CardList";
-export * from "@/page/ConfigPage";
-export * from "@/page/CardFormPage";
-export * from "@/page/DeckImportPage";
-export * from "@/page/CardViewPage";
