+++ src/pages/card-form/index.ts
@@ -1,0 +1,1 @@
+export { CardFormPage } from "./ui/CardFormPage";
