+++ src/pages/card-form/ui/CardFormPage.tsx
@@ -1,0 +1,3 @@
+import { CardFormContainer } from "@/features/card";
+
+export const CardFormPage = () => <CardFormContainer />;
