+++ src/pages/card-list/index.ts
@@ -1,0 +1,1 @@
+export { CardListPage } from "./ui/CardListPage";
