+++ src/pages/card-list/ui/CardListPage.tsx
@@ -1,0 +1,3 @@
+import { CardListContainer } from "@/features/card";
+
+export const CardListPage = () => <CardListContainer />;
