+++ src/pages/card-view/index.ts
@@ -1,0 +1,1 @@
+export { CardViewPage } from "./ui/CardViewPage";
