+++ src/pages/card-view/ui/CardViewPage.tsx
@@ -1,0 +1,3 @@
+import { CardViewContainer } from "@/features/card";
+
+export const CardViewPage = () => <CardViewContainer />;
