+++ src/pages/deck-form/index.ts
@@ -1,0 +1,1 @@
+export { DeckFormPage } from "./ui/DeckFormPage";
