+++ src/pages/deck-form/ui/DeckFormPage.tsx
@@ -1,0 +1,3 @@
+import { DeckFormContainer } from "@/features/deck";
+
+export const DeckFormPage = () => <DeckFormContainer />;
