+++ src/pages/deck-import/index.ts
@@ -1,0 +1,1 @@
+export { DeckImportPage } from "./ui/DeckImportPage";
