+++ src/pages/deck-import/ui/DeckImportPage.tsx
@@ -1,0 +1,3 @@
+import { DeckImportContainer } from "@/features/import";
+
+export const DeckImportPage = () => <DeckImportContainer />;
