+++ src/pages/deck-list/index.ts
@@ -1,0 +1,1 @@
+export { DeckListPage } from "./ui/DeckListPage";
