+++ src/pages/deck-list/ui/DeckListPage.tsx
@@ -1,0 +1,3 @@
+import { DeckListContainer } from "@/features/deck";
+
+export const DeckListPage = () => <DeckListContainer />;
