+++ src/pages/deck-start/index.ts
@@ -1,0 +1,1 @@
+export { DeckStartPage } from "./ui/DeckStartPage";
