+++ src/pages/deck-start/ui/DeckStartPage.tsx
@@ -1,0 +1,3 @@
+import { DeckStartContainer } from "@/features/study";
+
+export const DeckStartPage = () => <DeckStartContainer />;
