+++ src/pages/deck-swiper/index.ts
@@ -1,0 +1,1 @@
+export { DeckSwiperPage } from "./ui/DeckSwiperPage";
