+++ src/pages/deck-swiper/ui/DeckSwiperPage.tsx
@@ -1,0 +1,3 @@
+import { DeckSwiperContainer } from "@/features/study";
+
+export const DeckSwiperPage = () => <DeckSwiperContainer />;
