+++ src/pages/settings/ui/SettingsPage.tsx
@@ -1,0 +1,3 @@
+import { ConfigContainer } from "@/features/settings";
+
+export const SettingsPage = () => <ConfigContainer />;
